Move groups to interfaces
